requirements.txt
requests
python-dotenv
pytz
beautifulsoup4
ics
icalendar
playwright